recognition-api/requirements.txt
pytesseract

pyyaml
requests
scipy
tqdm
pandas
psutil
dill
py-cpuinfo

torch
torchvision

numpy
opencv-python-headless
pillow
matplotlib

fastapi
uvicorn
python-multipart